Expose isMethodCall matcher in flutter_test
