+++ packages/flutter/test/widgets/title_test.dart
@@ -6,8 +6,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter/services.dart';
 
-import '../services/message_codecs_utils.dart';
-
 void main() {
   testWidgets('toString control test', (WidgetTester tester) async {
     final Widget widget = new Title(
